backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
beautifulsoup4==4.12.2
markdown==3.5.1

# Vector Database and Embeddings
chromadb==0.4.15
sentence-transformers==2.2.2
huggingface-hub==0.17.3
transformers==4.35.2

# LLM Integration - Updated OpenAI version
openai==0.28.1
anthropic==0.7.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Production
gunicorn==21.2.0